Update WorkingFileForTesting.py
Changed loop and process of output to CSV to be able to check the values easier than printing to console. Each section can be written to dataframe, but it will not be the best method. Wrote up which sections seem to be better suited for JSON and inclusion to another record type. The organizing this back into a database will be the next step after confiming data formats and correct field pulls.