Update WorkingFileForTesting.py
Fleshed out more variable dataframes. Started playing with the unique keys necessary to match up the data in database form. I think I have isolated a way to separate out the data that is different for oil and gas wells. Formats for 02, 03, and 14 seem to be working correctly for the first 500 wells. Need to test more.