Update WorkingFileForTesting.py
Big update to the working file. All functions and libraries support keeping the data in bytes for as long as possible. All sections are loaded into to dataframes for format testing (QAQC). The current iteration looks at the first 50 wells in the file, but can be easily changed.

Advantages is that it runs faster as bytes and doesn't seem to bog down memory as much (this needs to be verified). Next steps are QAQC of all sections, which is made easier by the output of csv files. Then getting the remarks sections loaded as JSON records to their corresponding parent section.